CI: Run a different set of tests on macOS.

+++ scripts/ci/test.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 
-if test ! -e test-bundle/mono-libgdiplus.so; then
+if test x$CI_PLATFORM = xlinux -a ! -e test-bundle/mono-libgdiplus.so; then
 	wget -nv -O libgdiplus.zip "https://gitlab.winehq.org/mono/libgdiplus/-/jobs/artifacts/main/download?job=linux-build" || exit 1
 	unzip -q libgdiplus.zip || exit 1
 	cp usr/local/lib/libgdiplus.so test-bundle/mono-libgdiplus.so || exit 1
@@ -62,33 +62,51 @@ run_test mini-aot --mini --aot="mcpu=native"
 
 run_test runtime --runtime
 
-run_test verify --verify
-
-run_test mcs --mcs
-
-run_test mcs_errors --mcs_errors
-
-run_test symbolicate --symbolicate
-
-run_test profiler --profiler
-
-run_test csi --csi
-
 run_test aot-test --aot-test
 
-RETRIES=9 run_test System.Runtime.Caching --nunit net_4_x/tests/net_4_x_System.Runtime.Caching_test.dll
+if test x$CI_PLATFORM != xmac; then
+	for name in corlib System Mono.Posix ; do
+		run_test $name --nunit net_4_x/tests/net_4_x_${name}_test.dll
+	done
+fi
 
-for name in corlib System System.Xml Mono.Security System.Security System.Data Mono.Posix System.Web System.Web.Services System.Runtime.Serialization.Formatters.Soap System.Runtime.Remoting Cscompmgd Commons.Xml.Relaxng System.ServiceProcess I18N.CJK I18N.West I18N.MidEast I18N.Rare I18N.Other System.DirectoryServices Microsoft.Build.Engine Microsoft.Build.Framework Microsoft.Build.Tasks Microsoft.Build.Utilities Mono.C5 Mono.Options Mono.Tasklets System.Configuration System.Transactions System.Web.Extensions System.Core System.Drawing System.Windows.Forms Microsoft.VisualBasic System.Windows.Forms.DataVisualization System.Xml.Linq System.Data.DataSetExtensions System.Web.Abstractions System.Web.Routing System.Runtime.Serialization System.IdentityModel System.ServiceModel System.ServiceModel.Web System.ComponentModel.DataAnnotations Mono.CodeContracts System.Data.Services System.Web.DynamicData Mono.CSharp WindowsBase System.Numerics System.Runtime.DurableInstancing System.ServiceModel.Discovery System.Xaml System.Net.Http System.Net.Http.WebRequest System.Json System.Threading.Tasks.Dataflow Mono.Debugger.Soft Microsoft.Build System.IO.Compression Mono.Data.Sqlite System.Data.OracleClient Mono.Messaging Mono.Messaging.RabbitMQ WebMatrix.Data; do
+for name in Mono.Security System.Security System.Core Mono.Debugger.Soft; do
 	run_test $name --nunit net_4_x/tests/net_4_x_${name}_test.dll
 done
 
-run_test Microsoft.VisualBasic_vbtest --nunit net_4_x/tests/net_4_x_Microsoft.VisualBasic_vbtest.dll
-
-for name in corlib System System.Core System.Xml System.Runtime.CompilerServices.Unsafe System.Json System.Xml.Linq System.ComponentModel.Composition System.Drawing System.Security System.Runtime.Serialization System.Net.Http.UnitTests System.Net.Http.FunctionalTests Mono.Profiler.Log System.Data System.Numerics System.Threading.Tasks.Dataflow System.ComponentModel.Composition Microsoft.CSharp System.IO.Compression; do
+for name in corlib System System.Core; do
 	run_test ${name}-xunit --xunit net_4_x/tests/net_4_x_${name}_xunit-test.dll
 done
 
-MONO_TLS_PROVIDER=btls run_test corlib-btls --nunit net_4_x/tests/net_4_x_corlib_test.dll -include:X509Certificates
+if test x$CI_PLATFORM != xmac; then
+	MONO_TLS_PROVIDER=btls run_test corlib-btls --nunit net_4_x/tests/net_4_x_corlib_test.dll -include:X509Certificates
+fi
+
+if test x$CI_LIMITED = x; then
+	run_test verify --verify
+
+	run_test mcs --mcs
+
+	run_test mcs_errors --mcs_errors
+
+	run_test symbolicate --symbolicate
+
+	run_test profiler --profiler
+
+	run_test csi --csi
+
+	RETRIES=9 run_test System.Runtime.Caching --nunit net_4_x/tests/net_4_x_System.Runtime.Caching_test.dll
+
+	for name in System.Xml System.Data System.Web System.Web.Services System.Runtime.Serialization.Formatters.Soap System.Runtime.Remoting Cscompmgd Commons.Xml.Relaxng System.ServiceProcess I18N.CJK I18N.West I18N.MidEast I18N.Rare I18N.Other System.DirectoryServices Microsoft.Build.Engine Microsoft.Build.Framework Microsoft.Build.Tasks Microsoft.Build.Utilities Mono.C5 Mono.Options Mono.Tasklets System.Configuration System.Transactions System.Web.Extensions System.Drawing System.Windows.Forms Microsoft.VisualBasic System.Windows.Forms.DataVisualization System.Xml.Linq System.Data.DataSetExtensions System.Web.Abstractions System.Web.Routing System.Runtime.Serialization System.IdentityModel System.ServiceModel System.ServiceModel.Web System.ComponentModel.DataAnnotations Mono.CodeContracts System.Data.Services System.Web.DynamicData Mono.CSharp WindowsBase System.Numerics System.Runtime.DurableInstancing System.ServiceModel.Discovery System.Xaml System.Net.Http System.Net.Http.WebRequest System.Json System.Threading.Tasks.Dataflow Microsoft.Build System.IO.Compression Mono.Data.Sqlite System.Data.OracleClient Mono.Messaging Mono.Messaging.RabbitMQ WebMatrix.Data; do
+		run_test $name --nunit net_4_x/tests/net_4_x_${name}_test.dll
+	done
+
+	run_test Microsoft.VisualBasic_vbtest --nunit net_4_x/tests/net_4_x_Microsoft.VisualBasic_vbtest.dll
+
+	for name in System.Xml System.Runtime.CompilerServices.Unsafe System.Json System.Xml.Linq System.ComponentModel.Composition System.Drawing System.Security System.Runtime.Serialization System.Net.Http.UnitTests System.Net.Http.FunctionalTests Mono.Profiler.Log System.Data System.Numerics System.Threading.Tasks.Dataflow System.ComponentModel.Composition Microsoft.CSharp System.IO.Compression; do
+		run_test ${name}-xunit --xunit net_4_x/tests/net_4_x_${name}_xunit-test.dll
+	done
+fi
 
 touch "$TEST_RESULTS_DIR"/expected-failures.txt
